Middleware Multer + API Sauce route CRUD

+++ controllers/sauceController.js
@@ -1,3 +1,4 @@
+import fs from 'fs';
 import Sauce from '../models/sauce.js';
 
 export const getAll = (req, res, next) => {
@@ -27,20 +28,22 @@ export const getOne = (req, res, next) => {
 };
 
 export const create = (req, res, next) => {
-  const sauceRequest = JSON.parse(req.body.sauce);
+  const sauceObject = JSON.parse(req.body.sauce);
+  delete sauceObject._id;
+  delete sauceObject._userId;
+
   const sauce = new Sauce({
-    userId: sauceRequest.userId,
-    name: sauceRequest.name,
-    manufacturer: sauceRequest.manufacturer,
-    description: sauceRequest.description,
-    mainPepper: sauceRequest.mainPepper,
-    heat: sauceRequest.heat,
-    imageUrl: '',
+    ...sauceObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${
+      req.file.filename
+    }`,
     likes: 0,
     dislikes: 0,
     usersLiked: [],
     usersDisliked: [],
   });
+
   sauce
     .save()
     .then(() => {
@@ -56,37 +59,54 @@ export const create = (req, res, next) => {
 };
 
 export const modify = (req, res, next) => {
-  const sauce = new Sauce({
-    _id: req.params.id,
-    title: req.body.title,
-    description: req.body.description,
-    imageUrl: req.body.imageUrl,
-    price: req.body.price,
-    userId: req.body.userId,
-  });
-  Sauce.updateOne({ _id: req.params.id }, sauce)
-    .then(() => {
-      res.status(201).json({
-        message: 'Sauce updated successfully!',
-      });
+  const sauceObject = req.file
+    ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+
+  delete sauceObject._userId;
+
+  Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized' });
+      } else {
+        Sauce.updateOne(
+          { _id: req.params.id },
+          { ...sauceObject, _id: req.params.id }
+        )
+          .then(() =>
+            res.status(200).json({ message: 'Sauce modified succesfully!' })
+          )
+          .catch((error) => res.status(401).json({ error }));
+      }
     })
     .catch((error) => {
-      res.status(400).json({
-        error: error,
-      });
+      res.status(400).json({ error });
     });
 };
 
 export const deleteOne = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: 'Sauce deleted!',
-      });
+  Sauce.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Not authorized' });
+      } else {
+        const filename = sauce.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: 'Objet supprimé !' });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
     })
     .catch((error) => {
-      res.status(400).json({
-        error,
-      });
+      res.status(500).json({ error });
     });
 };
